refactor: 🔄 Rename for improved readability(#29)
