+++ src/pages/api/trpc/[trpc].ts
@@ -1,5 +1,5 @@
+import { createTRPCContext } from '@/server/trpc';
 import { appRouter } from '@/server/trpc/root';
-import { createTRPCContext } from '@/server/trpc/trpc';
 import { createNextApiHandler } from '@trpc/server/adapters/next';
 import { type NextApiRequest, type NextApiResponse } from 'next';
